// src/store/booking/types.ts
export interface SectionWithSeats extends SeatSectionBasic {
  selectedSeats?: SeatPosition[];
}
// ...
export type BookingState =
  | 'ADDONS_SELECTED'
  | 'CLASS_SELECTED'
  | 'ROUTE_SELECTED'
  | 'SEAT_SELECTED';

export interface BookingProps {
  bookingState: BookingState;
  codeDiscount?: string;
  flexiType?: FlexiType;
  fromStationId?: number;
  lineGroupCode?: string;
  lineNumber?: number;
  passengers?: PassengerData[];
  percentualDiscountIds?: number[];
  price?: number;
  discountAmount?: number;
  priceSource?: string;
  routeId?: RouteId;
  seatClass?: SeatClass;
  sections?: SectionWithSeats[];
  selectedAddons?: SelectedAddon[];
  tariffs?: Tariff[];
  toStationId?: number;
  type?: TicketType;
}